Reorganized structure of auditor.cleanup and auditor.demo
