+++ auditor.cleanup/src/main/java/io/github/tkaczenko/auditor/cleanup/internal/service/package-info.java
@@ -1,2 +1,2 @@
 /** Provides utility classes and services for cleaning up audited data. */
-package io.github.tkaczenko.auditor.cleanup.service;
+package io.github.tkaczenko.auditor.cleanup.internal.service;
