+++ demo/src/main/java/io/github/tkaczenko/auditor/demo/repository/AuditRecordRepository.java
@@ -1,7 +1,7 @@
 package io.github.tkaczenko.auditor.demo.repository;
 
 import io.github.tkaczenko.auditor.core.internal.repository.AuditRequestResponseRepository;
-import io.github.tkaczenko.auditor.demo.model.AuditRecord;
+import io.github.tkaczenko.auditor.demo.entity.AuditRecord;
 import org.springframework.stereotype.Repository;
 
 @Repository
