+++ demo/src/main/java/io/github/tkaczenko/auditor/demo/service/filter/DemoControllerAuditRequestFilter.java
@@ -1,9 +1,9 @@
 package io.github.tkaczenko.auditor.demo.service.filter;
 
-import io.github.tkaczenko.auditor.core.internal.service.AuditDateTimeProvider;
 import io.github.tkaczenko.auditor.core.internal.AuditFacade;
 import io.github.tkaczenko.auditor.core.internal.factory.body.BodyHttpReaderService;
 import io.github.tkaczenko.auditor.core.internal.factory.headers.HeadersHttpReaderService;
+import io.github.tkaczenko.auditor.core.internal.service.AuditDateTimeProvider;
 import io.github.tkaczenko.auditor.inbound.api.AuditInboundRequestFilter;
 import io.github.tkaczenko.auditor.inbound.internal.service.IpAddressSupplier;
 import org.springframework.stereotype.Component;
